Add package.json
PlatformIO thinks that this is a tool for some reason and starts looking for a package.json (I guess tools can be written in Node.js?). It goes and scans the whole repo and finds a package.json where it is not supposed to. So we need to create a dummy package.json.